416: composer fix
