+++ src/Payload/Context.php
@@ -42,7 +42,7 @@ class Context implements \Serializable
             "post" => $this->post,
         );
         
-        $objectHashes = \Rollbar\Utilities::GetObjectHashes();
+        $objectHashes = \Rollbar\Utilities::getObjectHashes();
         
         return $this->utilities->serializeForRollbar($result, null, $objectHashes);
     }
