+++ src/Payload/Request.php
@@ -150,7 +150,7 @@ class Request implements \Serializable
             $result[$key] = $val;
         }
         
-        $objectHashes = \Rollbar\Utilities::GetObjectHashes();
+        $objectHashes = \Rollbar\Utilities::getObjectHashes();
         
         return $this->utilities->serializeForRollbar($result, array_keys($this->extra), $objectHashes);
     }
