+++ src/Payload/Server.php
@@ -85,7 +85,7 @@ class Server implements \Serializable
             $result[$key] = $val;
         }
         
-        $objectHashes = \Rollbar\Utilities::GetObjectHashes();
+        $objectHashes = \Rollbar\Utilities::getObjectHashes();
         
         return $this->utilities->serializeForRollbar($result, array_keys($this->extra), $objectHashes);
     }
