+++ tests/TestHelpers/CycleCheck/ChildCycleCheck.php
@@ -1,0 +1,11 @@
+<?php namespace Rollbar\CycleCheck;
+
+class ChildCycleCheck
+{
+    public $parent;
+    
+    public function __construct($parent)
+    {
+        $this->parent = $parent;
+    }
+}
