+++ tests/TestHelpers/CycleCheck/ParentCycleCheck.php
@@ -1,0 +1,11 @@
+<?php namespace Rollbar\CycleCheck;
+
+class ParentCycleCheck
+{
+    public $child;
+    
+    public function __construct()
+    {
+        $this->child = new ChildCycleCheck($this);
+    }
+}
