+++ tests/TestHelpers/CycleCheck/ParentCycleCheckSerializable.php
@@ -1,0 +1,27 @@
+<?php namespace Rollbar\CycleCheck;
+
+class ParentCycleCheckSerializable implements \Serializable
+{
+    public $child;
+    
+    public function __construct()
+    {
+        $this->child = new ChildCycleCheck($this);
+    }
+    
+    public function serialize()
+    {
+        $objectHashes = \Rollbar\Utilities::GetObjectHashes();
+        return array(
+            "child" => \Rollbar\Utilities::serializeForRollbar(
+                $this->child,
+                null,
+                $objectHashes
+            )
+        );
+    }
+    
+    public function unserialize($serialized)
+    {
+    }
+}
